+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Changed
+
+- Refactor messages componet into livewire with `ig-message` event listener for dynamic messages.
+
 ## [0.10.2] - 2025-04-15
 
 ### Fixed
@@ -310,6 +316,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 ## [0.0.0] - 2024-09-12
 
+[Unreleased]: https://https://github.com/internetguru/laravel-common/compare/staging...dev
 [0.10.2]: https://https://github.com/internetguru/laravel-common/compare/v0.10.1...v0.10.2
 [0.10.1]: https://https://github.com/internetguru/laravel-common/compare/v0.10.0...v0.10.1
 [0.10.0]: https://https://github.com/internetguru/laravel-common/compare/v0.9.0...v0.10.0